# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type VaultDocumentStatus = typeof VaultDocumentStatus[keyof typeof VaultDocumentStatus];


export const VaultDocumentStatus = {
  Verified: 'Verified',
  Not_added: 'Not added',
  Processing: 'Processing',
} as const;

export type VaultDocumentColor = typeof VaultDocumentColor[keyof typeof VaultDocumentColor];


export const VaultDocumentColor = {
  aqua: 'aqua',
  gold: 'gold',
  navy: 'navy',
  coral: 'coral',
} as const;

export interface VaultDocument {
  id: string;
  type: string;
  label: string;
  status: VaultDocumentStatus;
  identifier: string;
  updated: string;
  icon: string;
  color: VaultDocumentColor;
  /** @nullable */
  imageData?: string | null;
  /** @nullable */
  contentType?: string | null;
}

export type VaultDocumentInputStatus = typeof VaultDocumentInputStatus[keyof typeof VaultDocumentInputStatus];


export const VaultDocumentInputStatus = {
  Verified: 'Verified',
  Not_added: 'Not added',
  Processing: 'Processing',
} as const;

export type VaultDocumentInputColor = typeof VaultDocumentInputColor[keyof typeof VaultDocumentInputColor];


export const VaultDocumentInputColor = {
  aqua: 'aqua',
  gold: 'gold',
  navy: 'navy',
  coral: 'coral',
} as const;

export interface VaultDocumentInput {
  /** @minLength 8 */
  ownerKey: string;
  type: string;
  label: string;
  status: VaultDocumentInputStatus;
  identifier: string;
  updated: string;
  icon: string;
  color: VaultDocumentInputColor;
  /** @nullable */
  imageData?: string | null;
  /** @nullable */
  contentType?: string | null;
}

export interface VaultDocumentDeleteInput {
  /** @minLength 8 */
  ownerKey: string;
}

export type ListVaultDocumentsParams = {
/**
 * @minLength 8
 */
ownerKey: string;
};