# lib/api-zod/src/generated/types/vaultDocument.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VaultDocumentColor } from './vaultDocumentColor';
import type { VaultDocumentStatus } from './vaultDocumentStatus';

export interface VaultDocument {
  id: string;
  type: string;
  label: string;
  status: VaultDocumentStatus;
  identifier: string;
  updated: string;
  icon: string;
  color: VaultDocumentColor;
  /** @nullable */
  imageData?: string | null;
  /** @nullable */
  contentType?: string | null;
}