# lib/api-zod/src/generated/types/vaultDocumentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VaultDocumentInputColor } from './vaultDocumentInputColor';
import type { VaultDocumentInputStatus } from './vaultDocumentInputStatus';

export interface VaultDocumentInput {
  /** @minLength 8 */
  ownerKey: string;
  type: string;
  label: string;
  status: VaultDocumentInputStatus;
  identifier: string;
  updated: string;
  icon: string;
  color: VaultDocumentInputColor;
  /** @nullable */
  imageData?: string | null;
  /** @nullable */
  contentType?: string | null;
}